requirements.txt
streamlit
openai
edge-tts
genanki
Pillow